Add CippTable, wrap CippDatatable
